Improvements for llm-to-code (#7)
* temporarily fix import issue

* remove all the pycache from git

* adapted overcooked_sim to work with the rest of the pipeline

* changed the assemble burgers json to deal with unstacking

* llm-to-code fixes

* Fix overcooked sim for merging to main

* Improve detection so that the project is runnable locally and from another directory

---------
