+++ environments/env_generator/examples/base_unstack.json
@@ -15,14 +15,14 @@
             "x": 0,
             "y": 1,
             "stack-level": 0,
-            "id": "a"
+            "id": "b"
         },
         {
             "name": "item",
             "x": 0,
             "y": 1,
             "stack-level": 1,
-            "id": "b"
+            "id": "a"
         }
     ],
     "players": [
@@ -37,7 +37,7 @@
         {
             "predicate": "has",
             "args": ["robot", "item"],
-            "ids": [1, "b"]
+            "ids": [1, "a"]
         }
     ]
 }
