+++ environments/env_generator/examples/composite_move_place.json
@@ -5,14 +5,13 @@
         {
             "name": "station",
             "x": 0,
-            "y": 1,
-            "id": "A"
+            "y": 1
         },
         {
             "name": "station",
             "x": 2,
             "y": 1,
-            "id": "B"
+            "id": "A"
         }
     ],
     "items": [
