+++ environments/env_generator/examples/composite_move_stack.json
@@ -5,14 +5,13 @@
         {
             "name": "station",
             "x": 0,
-            "y": 1,
-            "id": "A"
+            "y": 1
         },
         {
             "name": "station",
             "x": 2,
             "y": 1,
-            "id": "B"
+            "id": "A"
         }
     ],
     "items": [
@@ -21,14 +20,14 @@
             "x": 2,
             "y": 1,
             "stack-level": 0,
-            "id": "a"
+            "id": "b"
         },
         {
             "name": "item",
             "x": 0,
             "y": 0,
             "stack-level": 0,
-            "id": "b"
+            "id": "a"
         }
     ],
     "players": [
@@ -43,7 +42,7 @@
         {
             "predicate": "atop",
             "args": ["item", "item"],
-            "ids": ["b", "a"]
+            "ids": ["a", "b"]
         }
     ]
 }
