+++ environments/env_generator/examples/high_level_assemble_burgers.json
@@ -28,27 +28,27 @@
             "name": "patty",
             "x": 0,
             "y": 1,
-            "stack-level": 0,
+            "stack-level": 1,
             "predicates": ["iscooked", "iscookable"]
         },
         {
             "name": "patty",
             "x": 0,
             "y": 1,
-            "stack-level": 1,
+            "stack-level": 0,
             "predicates": ["iscooked", "iscookable"]
         },
         {
             "name": "topbun",
             "x": 1,
             "y": 1,
-            "stack-level": 0
+            "stack-level": 1
         },
         {
             "name": "topbun",
             "x": 1,
             "y": 1,
-            "stack-level": 1
+            "stack-level": 0
         },
         {
             "name": "bottombun",
